[FSSDK-10201] dir rename, testcase addition
