+++ lib/core/project_config/project_config_manager.ts
@@ -20,7 +20,7 @@ import { ERROR_MESSAGES } from '../../utils/enums';
 import { createOptimizelyConfig } from '../optimizely_config';
 import { OnReadyResult, OptimizelyConfig, DatafileManager } from '../../shared_types';
 import { ProjectConfig, toDatafile, tryCreatingProjectConfig } from '../project_config';
-import { scheduleMicrotaskOrTimeout } from '../../utils/microtasks';
+import { scheduleMicrotaskOrTimeout } from '../../utils/microtask';
 
 const logger = getLogger();
 const MODULE_NAME = 'PROJECT_CONFIG_MANAGER';
